Логика отображения ивентов перенесена в контроллер.
